% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{parameter_histograms_function}
\alias{parameter_histograms_function}
\title{Visualize sampled parameter distributions}
\usage{
parameter_histograms_function(mat)
}
\arguments{
\item{mat}{Parameter matrix produced by \link[=matrix_function]{matrix_function()}.}
}
\value{
A list of ggplot objects summarizing parameter draws.
}
\description{
Generates histograms for the Latin hypercube samples created by
\link[=matrix_function]{matrix_function()}, returning ggplot objects for reuse.
}
