% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSDplusplus-package.R
\docType{package}
\name{PSSDplusplus-package}
\alias{PSSDplusplus}
\alias{PSSDplusplus-package}
\title{PSSDplusplus: Probabilistic Microplastic Ecotoxicity Workflow (PSSD++)}
\description{
Implements the PSSD++ probabilistic ecotoxicological risk framework for
microplastics, as described in Coffin et al. (2026,
\doi{10.1016/j.jhazmat.2025.141021}). The package provides Monte Carlo
alignment of laboratory toxicity data to environmentally relevant metrics
(ERMs) -- food dilution and tissue translocation -- via Latin Hypercube
Sampling of particle-trait parameters. Aligned data are used to fit
probabilistic Species Sensitivity Distributions (pSSDs) and derive Predicted
No-Effect Concentrations (PNECs) across surface-water and sediment compartments
for freshwater and marine environments. Bundles the preprocessed ToMEx 2.0
microplastic ecotoxicity database (~13,000 data points) and a compact 10-species
subset for rapid testing. Supports parallel computation, deterministic caching,
and structured diagnostic output.
}
\references{
Coffin, S. et al. (2026). \emph{Journal of Hazardous Materials}
503, 141021. \doi{10.1016/j.jhazmat.2025.141021}
}
\author{
\strong{Maintainer}: Scott Coffin \email{scott.l.coffin@gmail.com} (\href{https://orcid.org/0000-0002-7035-1282}{ORCID})

}
