% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_ssdtools_version}
\alias{check_ssdtools_version}
\title{Ensure \code{ssdtools} is at the expected version (0.3.7)}
\usage{
check_ssdtools_version(required_version = "0.3.7")
}
\arguments{
\item{required_version}{Character version string. Defaults to \code{"0.3.7"}.}
}
\description{
This package relies on \code{ssdtools} 0.3.7 to reproduce published PSSD fits.
}
\examples{
\dontrun{
  check_ssdtools_version()
}
}
