% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSSDplusplus-package.R
\docType{package}
\name{PSSDplusplus-package}
\alias{PSSDplusplus}
\alias{PSSDplusplus-package}
\title{PSSDplusplus: Probabilistic Microplastic Ecotoxicity Workflow}
\description{
Implements the Monte Carlo alignment and PSSD++ analysis used for the ToMEx 2.0
microplastic ecotoxicity assessment. The package bundles helper functions,
plotting utilities, and workflow wrappers to reproduce the manuscript
thresholds with the bundled dataset.
}
\author{
\strong{Maintainer}: Scott Coffin \email{maintainer@example.com}

Other contributors:
\itemize{
  \item Contributors [contributor]
}

}
\keyword{internal}
